web/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-taggit==5.0.1
djangorestframework==3.15.1
gunicorn==21.2.0
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
